+++ src/com/aroundThirty/Controller/MainController.java
@@ -1,0 +1,7 @@
+package com.aroundThirty.Controller;
+
+
+public class MainController {
+    public MainController() {
+    }
+}
