+++ src/com/aroundThirty/Resource/BR.java
@@ -1,0 +1,8 @@
+package com.aroundThirty.Resource;
+
+import com.aroundThirty.Controller.EventController;
+import com.aroundThirty.model.ReportDto;
+
+public class BR {
+    public static ReportDto reportDto;
+}
