+++ src/com/aroundThirty/View/RightPanel.java
@@ -1,10 +1,8 @@
-package com.aroundThirty.mainPageGUI;
+package com.aroundThirty.View;
 
 import javax.swing.*;
 import javax.swing.border.LineBorder;
 import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 
 import static com.aroundThirty.Resource.FR.*;
